API updates in the cpp Rank methods
Remove the method `describeRank`.

The method originally give us a string
like "Jack-High Straight Flush". But because the table consumes too much
memory, and user can get more useful information using the
`describeSampleHand` method, so the `describeRank` API is dropped.

Update the return string in the `describeSampleHand` method

Originally the method returns a string like "9 9 8 7 6". Now the spaces
within the string is removed. So the current return values would be like
"99876".